requirements.txt
absl-py==2.1.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
attrs==24.2.0
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
imageio==2.36.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
MarkupSafe==3.0.1
matplotlib==3.9.2
mediapipe==0.10.15
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.4.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sounddevice==0.5.1
sympy==1.13.3
tifffile==2024.9.20
torch==2.4.1
torchsr==1.0.4
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
